# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.createSubtaskRequest:
    properties:
      collaborator_id:
        type: integer
      cr_id:
        type: integer
      done:
        type: boolean
      due_date:
        type: string
      progress:
        maximum: 100
        type: integer
      task_name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - cr_id
    - due_date
    - task_name
    type: object
...
  controllers.updateSubtaskRequest:
    properties:
      collaborator_id:
        type: integer
      cr_id:
        type: integer
      done:
        type: boolean
      due_date:
        type: string
      progress:
        maximum: 100
        type: integer
      task_name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - cr_id
    - due_date
    - task_name
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
  utils.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for MonitoringCR.
  title: MonitoringCR API
  version: "1.0"
paths:
...
  /api/subtasks:
    get:
      description: Get paginated list of subtasks, optionally filter by cr_id
      parameters:
      - description: Filter by CR ID
        in: query
        name: cr_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List Subtasks
      tags:
      - Subtask
    post:
      consumes:
      - application/json
      description: Create a new subtask.
      parameters:
      - description: Create Subtask payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controllers.createSubtaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create Subtask
      tags:
      - Subtask
  /api/subtasks/{id}:
    delete:
      description: Delete subtask by id.
      parameters:
      - description: Subtask ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete Subtask
      tags:
      - Subtask
    get:
      description: Retrieve one subtask by id.
      parameters:
      - description: Subtask ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get Subtask by ID
      tags:
      - Subtask
    put:
      consumes:
      - application/json
      description: Fully update subtask by id.
      parameters:
      - description: Subtask ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Subtask payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/controllers.updateSubtaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update Subtask
      tags:
      - Subtask
...
securityDefinitions:
  BearerAuth:
    description: Bearer token authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"